Challenge activities working.

+++ src/org/greengin/senseitweb/logic/project/challenge/ChallengeActivityActions.java
@@ -2,7 +2,6 @@ package org.greengin.senseitweb.logic.project.challenge;
 
 import java.util.Collection;
 import java.util.List;
-import java.util.Vector;
 
 import javax.persistence.TypedQuery;
 import javax.servlet.http.HttpServletRequest;
@@ -11,14 +10,15 @@ import org.greengin.senseitweb.entities.activities.challenge.ChallengeActivity;
 import org.greengin.senseitweb.entities.activities.challenge.ChallengeActivityStage;
 import org.greengin.senseitweb.entities.activities.challenge.ChallengeAnswer;
 import org.greengin.senseitweb.entities.activities.challenge.ChallengeField;
+import org.greengin.senseitweb.entities.activities.challenge.ChallengeOutcome;
 import org.greengin.senseitweb.entities.projects.Project;
 import org.greengin.senseitweb.logic.permissions.Role;
-import org.greengin.senseitweb.logic.project.AbstractActivityManager;
+import org.greengin.senseitweb.logic.project.AbstractActivityActions;
 import org.greengin.senseitweb.logic.voting.VoteCount;
 import org.greengin.senseitweb.logic.voting.VoteManager;
 import org.greengin.senseitweb.logic.voting.VoteRequest;
 
-public class ChallengeActivityManager extends AbstractActivityManager<ChallengeActivity> {
+public class ChallengeActivityActions extends AbstractActivityActions<ChallengeActivity> {
 
 	private static final String MY_ANSWERS_QUERY = String.format(
 			"SELECT a FROM %s a WHERE a.project = :project AND a.author = :author", ChallengeAnswer.class.getName());
@@ -30,7 +30,7 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 			"SELECT COUNT(a) AS N FROM %s a WHERE a.project = :project AND a.author = :author",
 			ChallengeAnswer.class.getName());
 
-	public ChallengeActivityManager(Long projectId, HttpServletRequest request) {
+	public ChallengeActivityActions(Long projectId, HttpServletRequest request) {
 		super(projectId, request, ChallengeActivity.class);
 	}
 
@@ -48,34 +48,31 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 				query.setParameter("author", user);
 			}
 			return query.getResultList();
-		} else {
-			return new Vector<ChallengeAnswer>();
-		}
+		} 
+		
+		return null;
 	}
 
 	public Collection<ChallengeAnswer> getAnswersForParticipant() {
 		if (hasAccess(Role.PROJECT_MEMBER)) {
 			return getAnswers(activity.getStage() == ChallengeActivityStage.PROPOSAL, true);
-		} else {
-			return new Vector<ChallengeAnswer>();
-		}
+		} 
+				
+		return null;
 	}
 
 	public Collection<ChallengeAnswer> getAnswersForAdmin() {
 		if (hasAccess(Role.PROJECT_ADMIN)) {
 			return getAnswers(false, false);
-		} else {
-			return new Vector<ChallengeAnswer>();
-		}
+		} 
+		
+		return null;
 	}
 
 	/** participant actions **/
 
 	public NewChallengeAnswerResponse createAnswer(ChallengeAnswerRequest answerData) {
-		NewChallengeAnswerResponse response = new NewChallengeAnswerResponse();
-		response.setNewAnswer(-1l);
-
-		if (hasAccess(Role.PROJECT_MEMBER)) {
+		if (hasAccess(Role.PROJECT_MEMBER) && activity.getStage() == ChallengeActivityStage.PROPOSAL) {
 			TypedQuery<Long> query = em.createQuery(ANSWER_COUNT_QUERY, Long.class);
 			query.setParameter("project", project);
 			query.setParameter("author", user);
@@ -90,45 +87,51 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 				em.persist(answer);
 				em.getTransaction().commit();
 
+				NewChallengeAnswerResponse response = new NewChallengeAnswerResponse();
 				response.setNewAnswer(answer.getId());
+				response.setAnswers(getAnswersForParticipant());
 			}
 		}
-
-		response.setAnswers(getAnswersForParticipant());
-		return response;
+		
+		return null;
 	}
 
 	public Collection<ChallengeAnswer> updateAnswer(Long answerId, ChallengeAnswerRequest answerData) {
-		if (hasAccess(Role.PROJECT_MEMBER)) {
+		if (hasAccess(Role.PROJECT_MEMBER) && activity.getStage() == ChallengeActivityStage.PROPOSAL) {
 			em.getTransaction().begin();
 			ChallengeAnswer answer = em.find(ChallengeAnswer.class, answerId);
 			answerData.update(answer);
 			em.getTransaction().commit();
+			return getAnswersForParticipant();
 		}
-
-		return getAnswersForParticipant();
+		
+		return null;
 	}
 
 	public Collection<ChallengeAnswer> deleteAnswer(Long answerId) {
-		if (hasAccess(Role.PROJECT_MEMBER)) {
-			em.getTransaction().begin();
+		if (hasAccess(Role.PROJECT_MEMBER) && activity.getStage() == ChallengeActivityStage.PROPOSAL) {
 			ChallengeAnswer answer = em.find(ChallengeAnswer.class, answerId);
+			em.getTransaction().begin();
+			if (activity.getOutcome().getSelectedAnswer() == answer) {
+				activity.getOutcome().setSelectedAnswer(null);
+			}
 			em.remove(answer);
 			em.getTransaction().commit();
+			return getAnswersForParticipant();
 		}
-
-		return getAnswersForParticipant();
+		
+		return null;
 	}
-	
+
 	public VoteCount vote(Long answerId, VoteRequest voteData) {
-		if (hasAccess(Role.PROJECT_MEMBER)) {
+		if (hasAccess(Role.PROJECT_MEMBER) && activity.getStage() == ChallengeActivityStage.VOTING) {
 			ChallengeAnswer answer = em.find(ChallengeAnswer.class, answerId);
 			if (answer != null && answer.getProject() == project) {
 				VoteManager voter = new VoteManager(em, user, answer);
 				return voter.vote(voteData);
 			}
 		}
-		
+
 		return null;
 	}
 
@@ -139,8 +142,10 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 			em.getTransaction().begin();
 			activity.setStage(stage);
 			em.getTransaction().commit();
+			return project;
 		}
-		return project;
+		
+		return null;
 	}
 
 	/** editor actions **/
@@ -150,9 +155,10 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 			em.getTransaction().begin();
 			activityData.update(activity);
 			em.getTransaction().commit();
+			return project;
 		}
 
-		return project;
+		return null;
 	}
 
 	public Project createField(ChallengeFieldRequest fieldData) {
@@ -162,8 +168,10 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 			fieldData.update(field);
 			activity.getFields().add(field);
 			em.getTransaction().commit();
+			return project;
 		}
-		return project;
+		
+		return null;
 	}
 
 	public Project updateField(Long fieldId, ChallengeFieldRequest fieldData) {
@@ -173,8 +181,11 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 			em.getTransaction().begin();
 			fieldData.update(field);
 			em.getTransaction().commit();
+			
+			return project;
 		}
-		return project;
+		
+		return null;
 	}
 
 	public Project deleteField(Long fieldId) {
@@ -186,6 +197,26 @@ public class ChallengeActivityManager extends AbstractActivityManager<ChallengeA
 			em.getTransaction().commit();
 		}
 		return project;
+	}
+
+	public ChallengeOutcome updateOutcome(ChallengeOutcomeRequest outcomeData) {
+		if (hasAccess(Role.PROJECT_ADMIN)) {
+			em.getTransaction().begin();
+			outcomeData.update(em, activity.getOutcome());
+			em.getTransaction().commit();
+			return activity.getOutcome();
+		}
+		
+		return null;
+	}
+
+	public ChallengeOutcome getOutcome() {
+		if (hasAccess(Role.PROJECT_ADMIN)
+				|| (hasAccess(Role.PROJECT_MEMBER) && activity.getStage() == ChallengeActivityStage.OUTCOME)) {
+			return activity.getOutcome();
+		} 
+		
+		return null;
 	}
 
 }
